Merged new conventions to leveled multidate heatplot, will save other changes for later
Need to consider additional changes to make plotting of missing values more robust
with more predefined ranges and increments, this and other changes will be saved
for post-presentation development.  Currently this will focus development on rerunning
template in the MET projects template directory as a shared resource, analyzing the
deep dive as the template use case.