Bug fixes and better error hanlding, including switch for hard / soft error on missing data
Warnings will be thrown if missing data unless told to fail.  Full preprocessing
runs through all options but will log the missing data with a failure
at the end of the script.  Using eval with embedded signal catch to
check for failures on MET / MET-tools-py.  Redirecting error to stdout
for convert_mpas to keep log of latlon file generation with other components.
Need to consider a recompile of the convert_mpas / singularity container.