Almost finished testing refactor of postprocess to DataFrames
Currently testing a version of the method used as a callable utility
with the parallelism running through Cylc a la the GridStat call itself.
This reduces several arguments simplifying the method and making a cleaner
method fit for single purpose.  Refactoring has led me to discovering a
bug in MET https://github.com/dtcenter/MET/issues/2730 that also exists
within GridStat.  Need to roll forward MET to 11.1.0 from 11.0.1 to check
if this is also fixed in GridStat or report bug.  Will resume testing
with MET rolled forward.