Templated full analysis with IVT / Precip treated with same conventions (avg. or acc. resp.)
Decided a consistent pipeline into the ensemble products for both variables would
be helpful for the analysis and streamlining.  Preparing now to complete pipeline to
gridstat for both of the mean / ensemble members directly.

Need to begin the creation / pipeline of ERA5 ground truth following while
running preprocessing tests on full data.