Sketched out main ideas for GenEnsProd, made regridding WRF default
Removed options to use native WRF grid with MET so that neighborhood methods can
be called with reference to the analysis field, precip / IVT, with respect
to a common default lat-lon grid.  Doubled resolution so that the lat-lon
intermediate grid is ~4km to try to get better performance out of
higher resolution models when verifying versus precip on 4km StageIV.

Main task now is to finish connecting the template for GenEnsProdMPAS
in Cylc with the driving script, smoothing conventions so that
precip / IVT are broadcast over tasks versus the same routines
within the driver, both using time interval quantities.