Removing traces, debugging appears to validate the IVT WRF method
Differences exist in the IVT / IWV calculations between NCL and Python
when comparing with legacy code, but the differences are not extreme.
In particular, differences with the IWV are smaller while IVT differences
are larger.  Checking over all codes, they appear to be working qualitatively
equivalently.  Will test versus additional production data on Narwhal.